DB relations and Mailable

+++ database/migrations/2022_05_10_100239_add_organization_id_to_users.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('organization_id')->nullable()->default(null)->after('remember_token');
+            $table->bigInteger('organization_id')->nullable()->default(null)->after('remember_token');
         });
     }
